adjusts

+++ futurelabs/lab/write.py
@@ -1,3 +1,4 @@
+import os
 from datetime import datetime
 
 import polars as pl
@@ -7,16 +8,18 @@ from sympy.stats.sampling.sample_numpy import numpy
 
 def write_chart_info(folder, chart):
     file_path = folder / "chart_info"
-    with open(file_path, "w") as file:
-        file.write(str(chart.value))
+
+    if not os.path.exists(file_path):
+        with open(file_path, "w") as file:
+            file.write(str(chart.value))
 
 def histogram(values, folder, chart):
+    write_chart_info(folder, chart)
 
     if len(values) > 0:
-        write_chart_info(folder, chart)
 
         df = pl.DataFrame({
-            "value": values
+            "value": [values]
         })
 
         timestamp = datetime.now().strftime("%Y_%m_%d_%H:%M:%S:%f")[:-3]
@@ -27,10 +30,6 @@ def histogram(values, folder, chart):
 def classification(data, folder):
     write_chart_info(folder, data["chart_type"])
 
-    # print(len(data["real_label"]))
-    # print(len(data["predicted_label"]))
-    # print(data["step"])
-
     df = pl.DataFrame(
         {
             "step": data["step"],
@@ -38,6 +37,8 @@ def classification(data, folder):
             "predicted_label": [data["predicted_label"]],
         }
     )
+
+    # print(df)
 
 
 
@@ -59,47 +60,6 @@ def scalar(values, step, folder, chart):
 
     file = folder / f"{step}.parquet"
     df.write_parquet(file)
-
-# def scalar(values, folder, chart):
-#
-#     write_chart_info(folder, chart)
-#
-#     steps = []
-#     metrics_columns = {}
-#
-#     # Iterar sobre cada dicionário na lista e organizar em colunas
-#     for entry in values:
-#         steps.append(entry["step"])  # Adiciona o valor de step à lista steps
-#         for key, value in entry["value"].items():
-#             if key not in metrics_columns:
-#                 metrics_columns[key] = []
-#             metrics_columns[key].append(value)
-#
-#     # Criar um dicionário organizado com os dados em colunas
-#     organized_data = {
-#         "step": steps,
-#         **metrics_columns  # Expande as colunas de metrics dinamicamente
-#     }
-#
-#     # Criar o DataFrame Polars
-#     df = pl.DataFrame(organized_data)
-#
-#     # Definir explicitamente os tipos das colunas
-#     df = df.with_columns([
-#         pl.col("step").cast(pl.Int32),  # Definir "step" como Int32
-#     ])
-#
-#     # Definir as colunas de métricas como Float32
-#     metric_columns = [pl.col(col).cast(pl.Float32) for col in metrics_columns.keys()]
-#     df = df.with_columns(metric_columns)
-#
-#     # Escrever o DataFrame em um arquivo Parquet
-#     timestamp = datetime.now().strftime("%Y_%m_%d_%H:%M:%S:%f")[:-3]
-#     file = folder / f"{timestamp}.parquet"
-#     df.write_parquet(file)
-#     # timestamp = datetime.now().strftime("%Y_%m_%d_%H:%M:%S:%f")[:-3]
-#     file = folder / f"{timestamp}.parquet"
-#     df.write_parquet(file)
 
 
 def audio(value, sr, step, folder, chart):
